+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: Include a fallback user for MSOnline, Security & Compliance and SfB
+
 ## 1.0.0 - 2021-02-09
 
 * Added: Initial version